% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_informations.R
\name{cor_test}
\alias{cor_test}
\title{remove the redundant variables}
\usage{
cor_test(df, cor.rate, info.vec, method = "spearman")
}
\arguments{
\item{df}{the dataframe}

\item{cor.rate}{the correlation threshold}

\item{info.vec}{the information vector}

\item{method}{there are three ways to compute correlation rate}

\item{num.vec}{must be numeric variables}
}
\description{
This function compute the correlation rate of two variables 
and remove the variables that its information value is less
}
\examples{
cor_test(p2p,0.7,con.vars,iv.list)
}

