% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_informations.R
\name{get_df_info}
\alias{get_df_info}
\title{get the information values of all colunms}
\usage{
get_df_info(df, targ, lowest, vars = names(df))
}
\arguments{
\item{df}{the dataframe}

\item{lowest}{the lowest samples that must have to make the iv robust}

\item{vars}{columns which need to compute;default:all columns will be used}

\item{target}{the target columns}
}
\description{
This function generatea all the information values of dataframe's columns
}
\examples{
get_df_info(df =p2p,targ='def',lowest=500)
}

