% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_informations.R
\name{get_IV}
\alias{get_IV}
\title{get a robust information value of a single variable}
\usage{
get_IV(prob, target, lowest = 500, miss = T)
}
\arguments{
\item{prob}{the variable that need to compute}

\item{target}{the label}

\item{lowest}{the lowest samples that must have to make the iv robust}
}
\description{
This function generatea  the variable's information values by a robust way
}
\examples{
get_IV(p2p[,1],p2p[,"def"],500)
}

