% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_informations.R
\name{get_pvalue}
\alias{get_pvalue}
\title{get the p values of single variable}
\usage{
get_pvalue(prob, target)
}
\arguments{
\item{prob}{the variable that need to compute}

\item{target}{the label}
}
\description{
This function using a single variable logistic regression to estimate the p value
}
\examples{
get_pvalue(p2p[,1],p2p[,"def"])
}

