% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{iv_filter}
\alias{iv_filter}
\title{filter the features}
\usage{
iv_filter(df, targ, iv.info, lowest = 500, method = "spearman",
  cor.rate = 0.7, dis.level = 0.02, con.level = 0.04)
}
\arguments{
\item{df}{the dataframe}

\item{targ}{the label variable}

\item{lowest}{the minimum seperation sample numbers to get a robust information}

\item{method}{the correlation test methods<spearson,pearson and kendall>}

\item{cor.rate}{the correlation threshold}

\item{dis.level}{the discrete variables' iv threshold}

\item{con.level}{the numeric variables' iv threshold,generally is bigger than discrete variables'}
}
\description{
using single variable's information to choose the features;
both discrete variables and numeric variables use the iv to determine their information;
using a iv threshold to choose those variables which are bigger than it;
and the numeric variables will be filtered by correlation test fothermore;
}
\examples{
iv_filter(p2p,targ="def",lowest=500,method="spearman",cor.rate=0.7,dis.level=0.02,con.level=0.04)
}

