% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_informations.R
\name{leave_0_point}
\alias{leave_0_point}
\title{transform the variable to a more robust stucture}
\usage{
leave_0_point(table.matrix, dis = F, lowest)
}
\arguments{
\item{table.matrix}{cross tabulation of two variables}

\item{dis}{boolean; if true,then reorder the table according to the odds of positive and negtive}

\item{lowest}{the smallest sample numbers that seperation must have}
}
\description{
This function generatea transform the variable to a more robust stucture
the function will remove those seperations that have no positive or negtive 
samples or the all samples numbers is too small
}
\examples{
leave_0_point(tables,dis=T,lowest=500)
}

