% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{miss_filter}
\alias{miss_filter}
\title{filter the features}
\usage{
miss_filter(df, iv.info, low.ml = 0.05, high.ml = 0.8, both.ml = 0.5,
  pvl = 0.001)
}
\arguments{
\item{df}{the dataframe}

\item{targ}{the label variable}

\item{lowest}{the minimum seperation sample numbers to get a robust information}

\item{method}{the correlation test methods<spearson,pearson and kendall>}

\item{cor.rate}{the correlation threshold}

\item{dis.level}{the discrete variables' iv threshold}

\item{con.level}{the numeric variables' iv threshold,generally is bigger than discrete variables'}
}
\description{
using single variable's missing information to choose the features;
both discrete variables and numeric variables use the iv to determine their information;
using a iv threshold to choose those variables which are bigger than it;
and the numeric variables will be filtered by correlation test fothermore;
}
\examples{
miss_filter(df,iv.info,low.ml=0.05,high.ml=0.8,both.ml=0.5,pvl=0.001)
}

