% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{rpart_filter}
\alias{rpart_filter}
\title{filter the features by decision tree}
\usage{
rpart_filter(df, targ, controls = rpart.control(minsplit = 100, maxdepth = 10,
  cp = 0, maxcompete = 1, maxsurrogate = 1, usesurrogate = 1, xval = 4,
  surrogatestyle = 0))
}
\arguments{
\item{df}{the dataframe}

\item{targ}{the label variable}

\item{controls}{the rpart tree's parameters}
}
\description{
using decision tree to choose the features
}
\examples{
rpart_filter(p2p,"def",controls=rpart.control(minsplit = 100,
                       maxdepth = 10, cp = 0, maxcompete = 1,
                      maxsurrogate = 1, usesurrogate = 1, 
                      xval =4,surrogatestyle = 0))
}

