% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{sbs_curve}
\alias{sbs_curve}
\title{plot the test-train ks}
\usage{
sbs_curve(fs.out)
}
\arguments{
\item{test.s}{the test ks serial}

\item{train.s}{the train ks serial}

\item{vars.s}{the variables serials}
}
\description{
plot the test-train ks to show how the variables' number influence ks
}
\examples{
sbs_curve(test.s,train.s,vars.s)
}

