% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{SFS}
\alias{SFS}
\title{using a sequence forward selection way to select variables}
\usage{
SFS(df, useful, candidate, targ, a, estimator)
}
\arguments{
\item{df}{the dataframe}

\item{useful}{the candidates variables}

\item{candidate}{the candidates variables}

\item{targ}{the label variable}

\item{a}{the threshold which determine whether the variable should be dropped}

\item{estimator}{the fit model to get some scoring}
}
\description{
using a sequence forward selection way to select variables
}
\examples{
SFS(p2p,useful,candidate,targ="def",a=0.013,estimator=xg_cv)
}

