% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{sig_filter}
\alias{sig_filter}
\title{filter the features}
\usage{
sig_filter(df, targ, iv.info, sig.info, lowest = 500, method = "spearman",
  cor.rate = 0.7, iv.level = 0.02, sig.level = 0.05)
}
\arguments{
\item{df}{the dataframe}

\item{targ}{the label variable}

\item{lowest}{the minimum seperation sample numbers to get a robust information}

\item{method}{the correlation test methods<spearson,pearson and kendall>}

\item{cor.rate}{the correlation threshold}

\item{iv.level}{the discrete variables' iv threshold}

\item{sig.level}{the numeric variables' p value threshold}
}
\description{
using single variable's information to choose the features;
discrete variables using iv as the critic and numeric variables use the significans,Pvalue as the evaluation critics;
using a iv threshold to choose those variables which are bigger than it;
and the numeric variables will be filtered by correlation test fothermore;
}
\examples{
sig_filter(p2p,targ="def",iv.info,lowest=500,method="spearman",cor.rate=0.7,dis.level=0.02,sig.level=0.05)
}

